+++ Assets/Domain/Inventory/ScriptedItem.cs
@@ -7,7 +7,7 @@ public class ScriptedItem : ScriptableObject
     public string ItemName = "New Item";
     public Sprite ItemIcon = null;
     public Rigidbody ItemSpawnObject = null;
-    public int PointCost = 1;
+    public float PointCost = 1;
     public AudioClip AudioWhenUsed = null;
     public GameObject AvailableSpawnPoints = null;
 }
